Moved to long integers.

+++ src/rshdf5.h
@@ -10,13 +10,13 @@
 namespace rshdf5 {
 
 ///Open the HDF5 file for writing
-int CreateFile(const std::string& name);
+long int CreateFile(const std::string& name);
 
 ///Add a dataset to the HDF5 file
-void AddChunkToFile(int file, std::complex<rsFloat> *data, unsigned int size, rsFloat time, rsFloat rate, rsFloat fullscale, unsigned int count);
+void AddChunkToFile(long int file, std::complex<rsFloat> *data, unsigned int size, rsFloat time, rsFloat rate, rsFloat fullscale, unsigned int count);
 
 ///Close the HDF5 file
- void CloseFile(int file);
+ void CloseFile(long int file);
 
 /// Read the pulse data from the specified file
  void ReadPulseData(const std::string &name, std::complex<rsFloat> **data, unsigned int &size, rsFloat &rate);
